# inline naive_kmer_hash and naive_miniser_hash