… test/include/seqan3/test/performance/naive_kmer_hash.hpp
… test/include/seqan3/test/performance/naive_minimiser_hash.hpp