The actual goal is to remove the usage of ranges::views::sliding. This
is easiest achieved by removing complete of `naive_kmer_hash` and `naive_minimizer_hash`

This is part of https://github.com/seqan/product_backlog/issues/124

-A life Without Range-v3-