@SGSSGene 2022-04-26
The benchmarks are very similar, for some reason there is more change in `seqan_kmer_hash_ungapped` (which I didn't change) than in `naive_kmer_hash`. Any Ideas why the `seqan_kmer_hash_ungapped` changed so much?

current Version:
```
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
seqan_kmer_hash_ungapped/1000/8          830 ns          830 ns       759898 Throughput[bp/s]=1.19651G/s
seqan_kmer_hash_ungapped/1000/30         728 ns          728 ns       948861 Throughput[bp/s]=1.33417G/s
seqan_kmer_hash_ungapped/50000/8       39925 ns        39925 ns        17327 Throughput[bp/s]=1.25218G/s
seqan_kmer_hash_ungapped/50000/30      36520 ns        36520 ns        19142 Throughput[bp/s]=1.36831G/s
seqan_kmer_hash_gapped/1000/8           5955 ns         5955 ns       116635 Throughput[bp/s]=166.745M/s
seqan_kmer_hash_gapped/1000/30         21967 ns        21967 ns        31821 Throughput[bp/s]=44.2024M/s
seqan_kmer_hash_gapped/50000/8        296380 ns       296381 ns         2350 Throughput[bp/s]=168.678M/s
seqan_kmer_hash_gapped/50000/30      1137668 ns      1137671 ns          616 Throughput[bp/s]=43.924M/s
naive_kmer_hash/1000/8                  2213 ns         2213 ns       316473 Throughput[bp/s]=448.798M/s
naive_kmer_hash/1000/30                 7338 ns         7338 ns        94570 Throughput[bp/s]=132.32M/s
naive_kmer_hash/50000/8               110636 ns       110635 ns         6308 Throughput[bp/s]=451.871M/s
naive_kmer_hash/50000/30              377885 ns       377886 ns         1853 Throughput[bp/s]=132.238M/s
```

PR version
```
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
seqan_kmer_hash_ungapped/1000/8          825 ns          825 ns       768144 Throughput[bp/s]=1.20313G/s
seqan_kmer_hash_ungapped/1000/30         879 ns          879 ns       790221 Throughput[bp/s]=1.10443G/s
seqan_kmer_hash_ungapped/50000/8       38567 ns        38567 ns        18220 Throughput[bp/s]=1.29627G/s
seqan_kmer_hash_ungapped/50000/30      41032 ns        41030 ns        17301 Throughput[bp/s]=1.2179G/s
seqan_kmer_hash_gapped/1000/8           6515 ns         6515 ns       106547 Throughput[bp/s]=152.407M/s
seqan_kmer_hash_gapped/1000/30         22329 ns        22329 ns        31242 Throughput[bp/s]=43.4863M/s
seqan_kmer_hash_gapped/50000/8        329464 ns       329465 ns         2120 Throughput[bp/s]=151.74M/s
seqan_kmer_hash_gapped/50000/30      1146840 ns      1146842 ns          610 Throughput[bp/s]=43.5727M/s
naive_kmer_hash/1000/8                  2163 ns         2163 ns       323162 Throughput[bp/s]=459.045M/s
naive_kmer_hash/1000/30                 7057 ns         7057 ns        97861 Throughput[bp/s]=137.595M/s
naive_kmer_hash/50000/8               107768 ns       107768 ns         6404 Throughput[bp/s]=463.894M/s
naive_kmer_hash/50000/30              363621 ns       363622 ns         1905 Throughput[bp/s]=137.426M/s
```