@eseiler 2022-04-26
There doesn't seem to be too much change?
You could increase the size (i.e. `1000`->`100'000` and `50'000` -> `5'000'000`) to try and see if this avoids cache artifacts.

We also do

https://github.com/seqan/seqan3/blob/e5c63f95004c67649ea0a6918348718852346e59/test/seqan3-test.cmake#L23-L26

The value could also be `64`....but I think increasing the benchmarked sequence sizes is enough.