@SGSSGene 2022-04-26
There was a second place where `ranges::views::sliding` was used. I updated this PR to also remove it from there.

Yes, you understand it correctly. This is the only place where `ranges::views::sliding` is being used (and the otherone that I just fixed ;-)).  We want to remove range-v3 sooner than later. Currently, no other functionality is using the sliding-view.

This also removes only the "naive" approaches. If you actually want to use the k-mer hashes you would use one of the other views provided by seqan. Those are a lot faster.

It still is possible to implement a sliding-view in the future.