@eseiler 2022-04-26
Just adding to what @SGSSGene said:

> I am currently mainly working for the FONDA project, but I have time so I am happy to review some PRs. 

If you ever are busy with FONDA, you can just remove your request and reassign the team.

> So you have to live with some possibly uninformed questions. ;)

Don't worry, there are no stupid questions and your questions are perfectly valid :)

> 
> **The way I understand it is this:** We want to get rid of range-v3 -> we want to get rid of `ranges::views::sliding` -> the only place where it is used is in the `naive_kmer_hash_fn` -> the `naive_kmer_hash_fn` is only used in a single benchmark, so we reimplement its use there and delete it.

👍 

> 
> **My questions:**
> 
> * Is this really the only place where `ranges::views::sliding` is used in SeqAn3? I have no idea, but it seems to me like this functionality might be actually used a lot. If that's the case, wouldn't it be better to reimplement the sliding view and leave everything here as is?
> * Are we really sure that we will never need the naive kmer hash function for any other benchmarks? We could also keep the `naive_kmer_hash_fn` and replace the sliding view there with our implementation.

It's the only place we use the range-v3 `views::sliding`. As we only us it for these two benchmarks, it would be quite a bit overhead to implement it as a view ... which we don't really have a use for.
We would implement it, if we have some proper use for it. Anyone having SeqAn3 as dependency and using the sliding view could still add range-v3 as dependency to their project.